Optimized code for works page
Added Sub Skeleton
